align tests structure with source code's
