+++ tests/unit/acquisition/multi_objective/test_partition.py
